Moved Claim back from common to aggregate service.
