+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/model/Claim.java
@@ -1,4 +1,6 @@
-package org.opentestsystem.rdw.reporting.common.model;
+package org.opentestsystem.rdw.olap.model;
+
+import org.opentestsystem.rdw.reporting.common.model.CodedEntity;
 
 /**
  * Represents a `scorable` claim per subject and assessment type
