+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/repository/ClaimRepository.java
@@ -1,6 +1,6 @@
 package org.opentestsystem.rdw.olap.repository;
 
-import org.opentestsystem.rdw.reporting.common.model.Claim;
+import org.opentestsystem.rdw.olap.model.Claim;
 
 import java.util.List;
