+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/repository/impl/JdbcClaimRepository.java
@@ -1,7 +1,7 @@
 package org.opentestsystem.rdw.olap.repository.impl;
 
 import org.opentestsystem.rdw.olap.model.AggregateServiceCache;
-import org.opentestsystem.rdw.reporting.common.model.Claim;
+import org.opentestsystem.rdw.olap.model.Claim;
 import org.opentestsystem.rdw.olap.repository.ClaimRepository;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cache.annotation.Cacheable;
