+++ aggregate-service/src/main/java/org/opentestsystem/rdw/olap/service/impl/ClaimReportQueryHandler.java
@@ -10,7 +10,7 @@ import org.opentestsystem.rdw.olap.repository.ClaimRepository;
 import org.opentestsystem.rdw.olap.service.AuthorizationService;
 import org.opentestsystem.rdw.reporting.common.model.AggregateQuery;
 import org.opentestsystem.rdw.reporting.common.model.AggregateReportType;
-import org.opentestsystem.rdw.reporting.common.model.Claim;
+import org.opentestsystem.rdw.olap.model.Claim;
 import org.opentestsystem.rdw.reporting.common.model.ClaimAggregateReportQuery;
 import org.opentestsystem.rdw.reporting.common.model.ClaimRow;
 import org.springframework.beans.factory.annotation.Autowired;
