+++ aggregate-service/src/test/java/org/opentestsystem/rdw/olap/service/impl/ClaimReportQueryHandlerTest.java
@@ -14,7 +14,7 @@ import org.opentestsystem.rdw.olap.repository.ReportRepository;
 import org.opentestsystem.rdw.olap.service.AggregateQueryConverter;
 import org.opentestsystem.rdw.olap.service.AggregateQueryHandler;
 import org.opentestsystem.rdw.olap.service.AuthorizationService;
-import org.opentestsystem.rdw.reporting.common.model.Claim;
+import org.opentestsystem.rdw.olap.model.Claim;
 import org.opentestsystem.rdw.reporting.common.model.ClaimAggregateReportQuery;
 import org.opentestsystem.rdw.reporting.common.model.ClaimRow;
 import org.opentestsystem.rdw.reporting.common.model.Dimension;
